Split up unit tests into multiple files
